Move SysmonGeneratorError to error module and remove unused error variants.
Unused error variants were hidden due to the impl for CheckedError
matching on all variants. All errors from this generator are expected to
be sqs_executor::errors::Recoverable::Persistent, so remove the match to
only return that value, which will surface unused SysmonGeneratorError
variants in the future.

The SysmonDecoderError is essentially removed and the impl for
PayloadDecoder just uses its own PayloadDecompressionError.
