+++ src/rust/generators/sysmon-generator/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod error;
 pub mod generator;
 pub mod metrics;
 pub mod models;
