+++ src/rust/generators/sysmon-generator/src/models/network/outbound.rs
@@ -21,7 +21,7 @@ use sysmon_parser::{
 };
 
 use crate::{
-    generator::SysmonGeneratorError,
+    error::SysmonGeneratorError,
     models::utc_to_epoch,
 };
